requirements.txt
black
isort
flake8
pylint
abstract-singleton
wheel
setuptools
build
twine
llama-cpp-python